Separate each route into individual files for scalability
