+++ src/components/AddExpensePage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AddExpensePage = () => (
+    <div>
+        This is from my add expense component
+    </div>
+);
+
+export default AddExpensePage;
