+++ src/components/EditExpensePage.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const EditExpensePage = () => (
+    <div>
+        This is from my edit expense component
+    </div>
+);
+
+export default EditExpensePage;
